third_party/blink/web_tests/platform/generic/fast/xpath/string-value-expected.txt

PASS doc.evaluate('.', doc, null, XPathResult.STRING_TYPE, null).stringValue is '0  1a-cdata-b2'
PASS doc.evaluate('.', doc.documentElement, null, XPathResult.STRING_TYPE, null).stringValue is '0  1a-cdata-b2'
PASS doc.evaluate('.', PI, null, XPathResult.STRING_TYPE, null).stringValue is '"Data"  '
PASS doc.evaluate('.', COMMENT, null, XPathResult.STRING_TYPE, null).stringValue is ' comment '
PASS doc.evaluate('.', CDATA, null, XPathResult.STRING_TYPE, null).stringValue is '-cdata-'
PASS doc.evaluate('.', TEXT, null, XPathResult.STRING_TYPE, null).stringValue is '0  '
PASS doc.evaluate('.', EMPTY, null, XPathResult.STRING_TYPE, null).stringValue is ''
PASS doc.evaluate('.', ATTR, null, XPathResult.STRING_TYPE, null).stringValue is '<&nbsp;>'
PASS successfullyParsed is true

TEST COMPLETE
